+++ lib/praxis/tasks/api_docs.rb
@@ -27,12 +27,13 @@ namespace :praxis do
     end
 
     desc "Run API Documentation Browser"
-    task :preview => [:install, 'docs:generate'] do
-      exec({'USER_DOCS_PATH' => File.join(Dir.pwd, 'docs')}, "#{path}/node_modules/.bin/grunt serve --gruntfile '#{path}/Gruntfile.js'")
+    task :preview, [:port] => [:install, :generate]  do |t, args|      
+      doc_port = args[:port] || '9090'
+      exec({'USER_DOCS_PATH' => File.join(Dir.pwd, 'docs'), 'DOC_PORT' => doc_port}, "#{path}/node_modules/.bin/grunt serve --gruntfile '#{path}/Gruntfile.js'")
     end
 
     desc "Build docs that can be shipped"
-    task :build => [:install, 'docs:generate'] do
+    task :build => [:install, :generate] do
       exec({'USER_DOCS_PATH' => File.join(Dir.pwd, 'docs')}, "#{path}/node_modules/.bin/grunt build --gruntfile '#{path}/Gruntfile.js'")
     end